Fix page-routes
