--- backend/README.md
@@ -1,11 +1,0 @@
-# Filtrering 
-
-# Filtrera highscores baserat på ordlängd:
-
-https://www.google.se/search?q=get+url+javascript+query&biw=1920&bih=955&ei=5MZIZJjbMcyOxc8P1pOGsAM&ved=0ahUKEwiY3YPn98b-AhVMR_EDHdaJATYQ4dUDCA4&uact=5&oq=get+url+javascript+query&gs_lcp=Cgxnd3Mtd2l6LXNlcnAQAzIGCAAQFhAeMgYIABAWEB4yBggAEBYQHjIGCAAQFhAeMgYIABAWEB4yBggAEBYQHjIGCAAQFhAeMggIABAIEB4QDTIICAAQCBAeEA0yCAgAEAgQHhANOgoIABBHENYEELADOgoIABCKBRCwAxBDOgUIABCABDoICAAQFhAeEA9KBAhBGABQpRRYjCpgkCtoAnABeACAAUWIAZEDkgEBNpgBAKABAcgBCsABAQ&sclient=gws-wiz-serp
-
-1. Skapa en <input type="range"> så att användaren kan välja filtrering baserat på ordlängd. Lägg till en eventListener till denna input som kör update-funktionen vid change. 
-
-2. Skapa en update-funktion i pagination .js. 
-denna funktion ska vara den centrala funktion som gör en ny request varje gång användaren gör en ändring (antingen klickar på ett pageNr eller ändrar type="range"). 
-
