+++ backend/src/db/schemas.js
@@ -1,7 +1,6 @@
 import mongoose from 'mongoose';
 
 const Highscore = mongoose.model('Highscore', {
-  // Mongoose tar första argumentet och lägger till ett S när/om collection skapas.
   name: String,
   time: Number,
   guesses: Array,
@@ -12,25 +11,3 @@ const Highscore = mongoose.model('Highscore', {
 });
 
 export { Highscore };
-
-// async function run() {
-//   const conn = await mongoose.connect('mongodb://127.0.0.1:27017/ordspel'); //anslutningssträng
-
-//   const highscore = new Highscore({
-//     name: 'Daniel',
-//     time: 12.34,
-//     guesses: ['SOL', 'ORD', 'RÖD'],
-//     settings: {
-//       wordLength: 'any',
-//       repeatingChars: true,
-//     },
-//   });
-
-//   await highscore.save();
-
-//   const highscores = await Highscore.find(); //Kommer ge oss alla våra highscores.
-//   console.log(highscores);
-
-//   conn.disconnect();
-// }
-// run();
