+++ backend/src/scss/main.scss
@@ -19,5 +19,6 @@ html {
 @import 'sections/WordRow.scss';
 @import 'sections/pagination.scss';
 @import 'sections/highscores.scss';
+@import 'sections/navbar.scss';
 
 @import 'misc.scss';
