+++ backend/src/scss/misc.scss
@@ -38,19 +38,6 @@ button {
   border: 1px solid #727171;
 }
 
-.navbar-custom {
-  height: 150px;
-  border-radius: 6px;
-  margin-bottom: 10px;
-  background-color: #f0f0f0;
-  font-size: 20px;
-  border: 1px solid #727171;
-  border-radius: 6px;
-  max-width: 1500px;
-  margin: auto;
-
-}
-
 .stdCtr {
   border: 1px solid #B9B9B9;
   border-radius: 6px;
