+++ backend/src/scss/sections/Entry.scss
@@ -6,10 +6,3 @@
     justify-content: center;
     margin: auto;
 }
-
-.settingsBtn {
-}
-
-.startBtn {
-
-}
