+++ backend/src/scss/sections/highscores.scss
@@ -13,7 +13,7 @@
   border: 1px solid black;
   border-radius: 8px;
   margin-bottom: 20px;
-  background-color: rgb(24, 24, 24);
+  background-color: $c-flatBlack;
   color: $c-lightYellow;
   min-width: 250px;
   max-width: 500px;
