+++ backend/src/scss/sections/navbar.scss
@@ -1,0 +1,11 @@
+.navbar-custom {
+    height: 150px;
+    border-radius: 6px;
+    margin-bottom: 10px;
+    background-color: wheat;
+    font-size: 20px;
+    border: 1px solid #727171;
+    border-radius: 6px;
+    max-width: 1500px;
+    margin: auto;
+  }
