+++ backend/src/scss/variables.scss
@@ -4,7 +4,7 @@ $c-lightPink: #fcb4b4;
 $c-lightGreen: #75F482;
 $c-mintGreen: #ccffd1;
 
-// $-flatBlack:
+$c-flatBlack: rgb(24, 24, 24);
 $c-darkGrey: rgb(53, 53, 53);
 $c-lightGrey: #b9b9b9;
