+++ backend/src/utils/randWord.js
@@ -1,7 +1,6 @@
 import wordList from './wordList.js';
 
 export default function randWord(settings) {
-  return 'BILL';
   let words = wordList;
   const repeatChars = settings.repeatingChars;
